change module hierarchy
Modules implementing the equation have been moved to a single module,
helpers functions moved out from their own modules and noon has been
inserted to the julian module.
