+++ src/lib.rs
@@ -24,17 +24,34 @@
 
 const DEGREE: f64 = std::f64::consts::PI / 180.;
 
-mod anomaly;
-mod center;
-mod declination;
 mod event;
-mod hourangle;
 mod julian;
-mod longitude;
-mod noon;
-mod perihelion;
-mod sunrise;
-mod transit;
+mod solar_equation;
 
 pub use crate::event::{DawnType, SolarEvent};
-pub use crate::sunrise::{sunrise_sunset, SolarDay};
+pub use crate::solar_equation::SolarDay;
+
+/// Calculates the sunrise and sunset times for the given location and date.
+///
+/// # Example
+///
+/// ```
+/// use sunrise::sunrise_sunset;
+///
+/// // Calculate times for January 1, 2016 in Toronto
+/// let (sunrise, sunset) = sunrise_sunset(43.6532, -79.3832, 2016, 1, 1);
+/// ```
+pub fn sunrise_sunset(
+    latitude: f64,
+    longitude: f64,
+    year: i32,
+    month: u32,
+    day: u32,
+) -> (i64, i64) {
+    let solar_day = SolarDay::new(latitude, longitude, year, month, day);
+
+    (
+        solar_day.event_time(SolarEvent::Sunrise),
+        solar_day.event_time(SolarEvent::Sunset),
+    )
+}
