+++ src/solar_equation/anomaly.rs
@@ -28,7 +28,7 @@ const J2000: f64 = 2451545.;
 
 /// Calculates the angle of the sun relative to the earth for the specified
 /// Julian day.
-pub fn solar_mean_anomaly(day: f64) -> f64 {
+pub(crate) fn solar_mean_anomaly(day: f64) -> f64 {
     let v = ((357.5291 + 0.98560028 * (day - J2000)) * DEGREE) % (2. * PI);
 
     if v < 0. {
