+++ src/solar_equation/longitude.rs
@@ -22,10 +22,10 @@
 
 use std::f64::consts::PI;
 
-use crate::perihelion;
+use super::perihelion;
 
 /// Calculates the angular distance of the earth along the ecliptic.
-pub fn ecliptic_longitude(solar_anomaly: f64, equation_of_center: f64, day: f64) -> f64 {
+pub(crate) fn ecliptic_longitude(solar_anomaly: f64, equation_of_center: f64, day: f64) -> f64 {
     (solar_anomaly
         + equation_of_center
         + perihelion::argument_of_perihelion(day) % (2. * PI)
